Refactor AggregationHandle to avoid cyclic dependency when using HashTable.
